readme add examples

+++ src/main/kotlin/com/github/navyazaveri/dynamik/stdlib/containers/List.kt
@@ -40,7 +40,7 @@ class ListInstance : ContainerInstance() {
         env.defineFunction("add", NativeCallable("list.add", 1) { _list.add(it[0]) })
         env.defineFunction("get", NativeCallable("list.get", 1) { _list[(it[0] as Double).toInt()] })
         env.defineFunction("contains", NativeCallable("list.contains", 1) { _list.contains(it[0]) })
-        env.defineFunction("len", NativeCallable("list.len", 0) { _list.size })
+        env.defineFunction("len", NativeCallable("list.len", 0) { _list.size.toDouble() })
     }
 
     fun contains(thing: Any): Boolean {
